# project-management-system/PMS-Swagger- Documentation.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SriramB/ProjectManagementSystem/1.0.0
  - description: URL for PMS
    url: http://localhost:8080/api/v1
info:
  description: This API is PMS
  version: "1.0.0"
  title: PMS
  contact:
    email: suba@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /user/addUser:
    post:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: Add a new  User Record
      operationId: addUser
      description: |
        Add a new User Record
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UserList'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserListContent'  
  /user/getUser/{userId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for particular User
      operationId: getUser
      description: |
        Search for particular User
      parameters:
        - in: path
          name: userId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: One User
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input            
  /user/getUsers:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for All Users
      operationId: geUsers
      description: |
        Search for All Users
      responses:
        '200':
          description: List of Users
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UserList'
        '400':
          description: Invalid Input
  /user/getUsersByRole/{UserRoles}:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
        - bearerAuth: []
      summary: Search for All Users By Role
      operationId: getUsersByRole
      description: |
        Search for All Users By Role
      parameters:
        - in: path
          name: UserRoles
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: List of Users for a Particular Role
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UserListRole'
        '400':
          description: Invalid Input
  /user/getUsersByProject/{projectId}:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
        - bearerAuth: []
      summary: Search for All Users By Project Id
      operationId: getUsersByProject
      description: |
        Search for All Users By Project Id
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: List of Users for a Particular Project
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UserListRole'
        '400':
          description: Invalid Input                    
  /user/deleteUser/{userId}:
    delete:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: delete an existing User
      operationId: deleteUser
      description: |
        delete an existing Role
      parameters:
        - in: path
          name: userId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedUser'
        '400':
          description: Invalid Input  
  /user/updateUser/{userId}:
    patch:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: update existing User
      operationId: updateUser
      description: |
        update existing User
      parameters:
        - in: path
          name: userId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/UpdatedUser'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserListContent'    
...
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - UserDetail
      properties:
        UserDetail:
            $ref: '#/components/schemas/UserListContent'
    DeletedUser:
      type: object
      required:
        - Deleted
      properties:
        Deleted:
            $ref: '#/components/schemas/UserListContent'
    UpdatedUser:
      type: object
      required:
        - Deleted
      properties:
        Updated:
            $ref: '#/components/schemas/UserListContent'
    UserList:
      type: object
      required:
        - Users
      properties:
        Data:
          items:
              $ref: '#/components/schemas/UserListContent'
    UserListRole:
      type: object
      required:
        - Users
      properties:
        Users:
          items:
              $ref: '#/components/schemas/UserListContent'
    UserListContent:
      type: object
      required:
        - username
        - password
        - firstName
        - lastName
        - projectId
        - mobileNumber
        - emailId
        - dateOfJoining
        - dateOfResigning
        - userRoles
      properties:
        userId:
          type: string
          example: U11
        username:
          type: string
          example: john2k
        password:
          type: string
          example: John@2020
        firstName:
          type: string
          example: john
        lastName:
          type: string
          example: bk
        projectId:
          type: string 
          example: P1
        mobileNumber:
          type: string
          example: 9806546632
        emailId:
          type: string
          example: jogn@gmail.com
        createdBy :
          type: string
          example: "U14"
        LastUpdatedBy:
          type: string
          example: "U15"
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
        dateOfJoining:
          type: object
          example: 02-05-2020
        dateOfResigning:
          type: object
          example: 04-07-2039
        userRoles:
          type: object
          example: ["developer","tester"]
        userStatus:
          type: string
          example: active
...